requirements.txt
selene==2.0.0rc3post2
pytest~=7.4.0
selenium~=4.10.0
allure-pytest
python-dotenv~=1.0.0
